# Optimize Windows DX12 frame latency via DXGI waitable swap chain