This PR attempts to reduce frame latency for Windows DX12 by leveraging DXGI's waitable swap chain option.  Note this is not applicable to Vulkan.  I found some interesting articles and sample code on the subject:

1. https://www.intel.com/content/www/us/en/developer/articles/code-sample/sample-application-for-direct3d-12-flip-model-swap-chains.html
2. https://jackmin.home.blog/2018/12/14/swapchains-present-and-present-latency/

This concept was very easy to implement for RBDoom3BFG and leads to some pretty good results on my AMD 6600XT card driving a 60 HZ monitor, with VSync set on:

1. Baseline queue submit to present time latency = **~53ms @ 60Hz** for GPU triple buffering (**NUM_FRAME_DATA** = 3) coupled with 3 swap chain buffers (**swapChainBufferCount** = 3)
2. Using waitable swap chains, you can program the desired latency at runtime to be from 1 to 3 frames.  Using the same GPU triple buffering with 3 swap chain buffers, the resulting latency is about **36 ms @ 60Hz for 2 frames** of latency, and **24 ms @ 60Hz for 1 frame** of latency.  These numbers would be lower if using a 120Hz or 144Hz monitor.
3. Note these vsync-on latency values pale in comparison to using **vsync off** (i.e. tearing enabled) with latency numbers in the **3-6 ms** range.  However, the trade-off is reduced image quality.

The above latency numbers were observed in windowed-mode using PresentMon in conjunction with the Optick profiler improvements from my previous pull request #780.  Borderless fullscreen mode gave similar results.

The tradeoff for lower latency is reduced GPU to CPU overlap and FPS throughput.  However, if using a powerful GPU this may not be very noticeable.  Even with my relatively low-powered 6600 XT, I can easily drive the game in the headquarters hallway scene at 60Hz with the DXGI waitable object set to 1 or 2 frames of latency.  However, I am not sure how things would perform with only 1 frame of latency (effectively no CPU/GPU parallelism) during heavy action sequences.  For this reason I am recommending a frame latency of 2 to achieve some CPU/GPU parallelism, mirroring the  recommendations in article 1 above.  In my experiments with RBDoom3BFG, I still found it important to keep NUM_FRAME_DATA = 3 for best performance, coupled with the waitable swap chain set to 2 frames of latency.

I have defined a new cvar, r_maxFrameLatency with default value of 2, to allow simple changes and experimentation.  Permitted values are 0, 1, 2, and 3.  The value of 0 turns off the feature.  Values 1 and 2 are useful and correspond to the number of queued back buffers permitted in the swap chain.  A value of 3 means using the full set of 3 swap chain buffers (same latency as off).  Note this cvar cannot be changed on the fly (CVAR_INIT), and must be set up when the swap chain is first initialized. To change it you must use the autoexec.cfg file or enter `seta r_maxFrameLatency <x>` in the console and restart the game.

Here are a couple of Optick screen grabs to show the results.  You can see that the VSync/Present queue is now labeled with the FrameID, allowing direct inspection of latency:

Capture 1 showing ~52 msec of latency (60Hz, vsync on, r_maxFrameLatency = 0, same as current app):

![Windows DX12 Windowed Present Buffers=3](https://github.com/RobertBeckebans/RBDOOM-3-BFG/assets/82544213/6abbc419-5822-451e-b947-128b45b18fd1)

Capture 2 showing ~36 msec of latency (60Hz, vsync on, r_maxFrameLatency = 2):

![Windows DX12 Windowed Present Buffers=3,FL=2](https://github.com/RobertBeckebans/RBDOOM-3-BFG/assets/82544213/bb93cd17-9e2d-410d-a443-d162ac1cf56b)

Capture 3 showing ~24 msec of latency (60Hz, vsync on, r_maxFrameLatency = 1):

![Windows DX12 Windowed Present Buffers=3,FL=1](https://github.com/RobertBeckebans/RBDOOM-3-BFG/assets/82544213/eb109716-249f-4acf-8851-802f9d516bbd)

Capture 4 showing the difference between the swapchain waitable object or "DX12_Sync1", and GPU triple buffering sync or "DX12_Sync3" (60 Hz, vsync on, r_maxFrameLatency = 3).  The "DX12_Sync1" point aligns with the start of a new present frame, and "DX12_Sync3" aligns with completion of the N-2 frame's GPU frame.  Depending on the settings and game circumstance, both may be observable as seen below:

![Windows DX12 Windowed Sync1+Sync3, FL=2](https://github.com/RobertBeckebans/RBDOOM-3-BFG/assets/82544213/4966eaa6-3b47-4d2d-a4d8-ad16f3404de3)