@Calinou 2023-05-27
Can a double-buffered V-Sync option be added to minimize input lag? Depending on your settings, you may be able to reach the maximum FPS at all times.