@SRSaunders 2023-05-28
The short answer is yes, the capability is already there.  There are two mechanisms to deliver double-buffering (or equivalent latency):

1. Set `NUM_FRAME_DATA = 2` (in precompiled.h) and recompile the current code base (i.e. without this PR).  This will reduce latency to about **34 msec @ 60Hz**, but your FPS results will depend entirely on how fast a GPU you have.  With only 2 GPU frame data slots, CPU/GPU processing is serialized and throughput is reduced.  For instance, with my RX6600 XT, I can no longer achieve 60 FPS (more like 45 FPS) in the benchmark headquarters hallway scene, whereas it easily surpasses 60 FPS (more like 200+ FPS unlocked with vsync off) when NUM_FRAME_DATA = 3 (i.e. triple buffering, with better CPU/GPU parallelism).
2. Alternatively, with this PR (if merged), you can leave `NUM_FRAME_DATA = 3` for higher throughput and FPS, but set `r_maxFrameLatency = 2` (the default) to achieve a similar latency of about **36 msec @ 60Hz**.  The end-to-end latency result will be the same as option 1, but with a higher frame rate.  That is the point of this PR.

Note with this PR (if merged) you can push the latency even lower by setting `r_maxFrameLatency = 1` resulting in a latency of **20-24 msec @ 60 Hz**.  However, this also eliminates CPU/GPU parallelism, so FPS performance will be lower and the ability to handle processing spikes without FPS drops will be reduced.  Again your mileage may vary based on the strength of your GPU.

If you want to experiment, I suggest you try out this PR and provide feedback on your results.  I would be very interested in how this works for other setups.  Alternatively, you could try out the Intel sample application in link 1 of my PR write-up above.  I found this interactive latency demo very help in understanding the tradeoffs.